Add a check so that installation occurs only in Termux
It checks whether /data/data/com.termux/files directory exists or not.

I know this is not the best way to check whether a user uses Termux or
not.  You could just make the same directory (if you have access), and
that would make the script thinks that you are using Termux.  But I
think this will work properly for now, unless you interfere.