.path: Do not remove ~/storage directory anymore
This update is really important; If you previously had stored your
important files in ~/storage directory and tried to install
termux_dotfiles, your files (not including linked files) may have been
removed by now.  Thus, Avoid using the previous version.

The ~/storage directory was intended to be removed because it is not a
part of termux_dotfiles.  It is an extraneous directory created by
termux built-in command: termux-setup-storage.  The purpose of running
the command is to create the EXTERNAL_STORAGE directory, as we don't
have access to create that directory within micro-sd card without root
access.