Fix dependencies installer script
The old installer script is apparently checking packages by the command
name, whether a package is installed or not.  That was a wrong check.
Thus, a package should be checked by its package name instead.  Suppose
if a package and its command name were different, then the installer
would install that package every single time.

Although the script checks for a package with the same name as the
package name does, the Bash built-in command "command -V" treats the
same name to be a command instead.  As a result, the check always fails
since there will be no command to be found.

So, to properly install the dependencies, all of the package names will
be passed to the pkg command as arguments.  Since it performs update and
install out of the box.  So, no more package checking.