Remove functions: cld and celd
Those 2 functions are now removed because of options handling issue.  As
it is impossible to handle all ls' valid options unless tested by ls
itself.

Example of running cld like this:
   cld -w 72 $INTERNAL/Download;

Would ruin the cld function, if 72 were a directory.  Because cld treats
each argument independently.  For this to be fixed up, cld has to test
the whole arguments with ls command whether there is an invalid option
or not.  Now, how would cld know those options have to be tested at
once, and not one at a time, without hard coding the ls options.

As known, "-w" option in ls required an argument, so 72 here cannot be
the argument for a directory that will be cd-ed, and thus should be
for "-w" option.  Meanwhile, If I were hard coding all of the options
now, it is possible in the future new options would be added or old
options removed.  If that's the case it will be cumbersome and hard to
maintain those 2 functions.  So just remove those functions altogether.