.exports: Change HISTCONTROL to erasedups
The previous value of HISTCONTROL environment variable was ignoreboth;
Which causes lines that begin with a space character as well as lines
matching the previous history entry to not be saved in the history list.
Now I change it to erasedups, which is the same, except that this causes
all the previous lines matching the current line to be removed from the
history list before that line is saved.