Merge separated functions of .cronjob commmands
Formerly, the .cronjob commands was implemented as functions for each
.cronjob object.  There are 6 objects in total by now.  That means,
there are also 6 functions that take care of those objects respectively.
It's getting hard to maintain it that way.  Moreover, the .cronjob
objects can be really much.  So, to solve this, I merged those functions
by command names; Because command names are rarely to be numerous.